Added contact add forms
	althought the form is rendering correctly, there is still a lot
of functionality to be added and tested. contact add view inside contact
app will redirect to the newly created contact after creation.

	future work:
	come up with fancy ways to display contacts, and share contacts,
as well as fill contacts. One idea is to allow new contacts to fill
their information based on their facebook profile. Ofcourse not all of
the information would be fillable but at the very least some would.